Update copyright headers in Java files to reference all contributors
The intent is that I expect git-history to track the precise
contributions from here on, and I hope this will help the project be
open to them.

I reached out to Andrew asking if he had any preference on how to host
this, before forking and extending, and his reply was:

  "That's all fine, I'm happy for you to do what's easiest for you."

The file lib/CHANGES.TXT contains the best known info about
contributions leading up to the 0.1.12 release. (Again, because the
repo hosted on java.net went away.)