Bump version to 0.2.0
This is a somewhat symbolic checkpoint (it's not intended for release).

At this point the API should be unchanged. However the minor number
bump helps signify that the library will no longer support JDK-7 and
earlier.

From here the API may be extended, although it should strive to maintain
backward compatibly throughout the 0.2 series.