Introduce method steps() and use to simplify for-loops
It would be more natural to implement stepsIterator() in terms of
steps() (or replace it completely) but this is not possible in the
interface, while maintaining API backwards compatibility.

As a compromise, we implement it like this in the 3 concrete classes
(where we can be sure that steps() does not fall back to the default
implementation).

At the same time, we change steps() to return a 'snapshot' copy, rather
than returning an UnmodifyiableList. This is cleaner to reason about.
There is now no chance of the iterator throwing if some code happens to
call add(WizardStep) in the middle of iterating. Although unlikely,
such possibilities are now completely ruled out. As a minor bonus, the
implementation and return type is consistent with the default impl.