Rename subproject 'lib' > 'wizard'
I'd misunderstood the Gradle init script, thinking this was a
delineation akin to a 'src' dir. However publication seems to want to
use this to determine the name of the '.jar' file. (By convention, but
those are often deeply baked in, and best not meddled with.)

About here I also considered merging the 2-gradle projects into 1, but
in the end I decided to keep low-utility outer project:

 - So that the git-hub project can be named 'wizard-framework'
   differently from the published artefact 'wizard'.

 - In case someone wants to add a separate demo-app in the future.