Remove the ant build scripts
It would be confusing to have both ant and gradle build definitions.

I think the main thing not covered by gradle is the generation of the
www page (for viewing on java.net which went away anyway?) so that will
just lurk around for viewing from source, with a few placeholder
variables.